Besides some contextual conflicts, resolve problems as follows:

* Update PrettyPrinter.h changes for r339132's change from bool to
  unsigned.

* Update for r342311's renames: getLocStart -> getBeginLoc and
  getLocEnd -> getEndLoc.

* Update StmtPrinter.cpp's changes for r342311's replacement of "\n"
  with NL and associated changes to function parameters.

* Update tests of reductions on complex types to use -latomic when
  available because r337722 removes automatic use of -latomic.

+++ lldb/source/Expression/DWARFExpression.cpp
@@ -16,12 +16,12 @@
 #include <vector>
 
 #include "lldb/Core/Module.h"
-#include "lldb/Core/RegisterValue.h"
-#include "lldb/Core/Scalar.h"
 #include "lldb/Core/Value.h"
 #include "lldb/Core/dwarf.h"
 #include "lldb/Utility/DataEncoder.h"
 #include "lldb/Utility/Log.h"
+#include "lldb/Utility/RegisterValue.h"
+#include "lldb/Utility/Scalar.h"
 #include "lldb/Utility/StreamString.h"
 #include "lldb/Utility/VMRange.h"
 
@@ -2382,7 +2382,7 @@ bool DWARFExpression::Evaluate(
     case DW_OP_lit29:
     case DW_OP_lit30:
     case DW_OP_lit31:
-      stack.push_back(Scalar(op - DW_OP_lit0));
+      stack.push_back(Scalar((uint64_t)(op - DW_OP_lit0)));
       break;
 
     //----------------------------------------------------------------------
